fix according to comments.
