+++ pkg/client/metadata.go
@@ -1,0 +1,15 @@
+package client
+
+// DubboMetadata dubbo metadata, api config
+type DubboMetadata struct {
+	ApplicationName      string   `yaml:"application_name" json:"application_name" mapstructure:"application_name"`
+	Group                string   `yaml:"group" json:"group" mapstructure:"group"`
+	Version              string   `yaml:"version" json:"version" mapstructure:"version"`
+	Interface            string   `yaml:"interface" json:"interface" mapstructure:"interface"`
+	Method               string   `yaml:"method" json:"method" mapstructure:"method"`
+	Types                []string `yaml:"types" json:"types" mapstructure:"types"`
+	Retries              string   `yaml:"retries"  json:"retries,omitempty" property:"retries"`
+	ClusterName          string   `yaml:"cluster_name"  json:"cluster_name,omitempty" property:"cluster_name"`
+	ProtocolTypeStr      string   `yaml:"protocol_type"  json:"protocol_type,omitempty" property:"protocol_type"`
+	SerializationTypeStr string   `yaml:"serialization_type"  json:"serialization_type,omitempty" property:"serialization_type"`
+}
